Created subclasses of Piece
- Made Piece.cs an abstract class and created subclasses of Piece to improve extendability of the code and make it easier to add new types of piece.
