+++ Tatedrez-Game/Assets/Prefabs/Piece.prefab
@@ -12,7 +12,6 @@ GameObject:
   - component: {fileID: 5001380633490382160}
   - component: {fileID: 5001380633490382161}
   - component: {fileID: 5001380633490382163}
-  - component: {fileID: 5001380633490382158}
   m_Layer: 5
   m_Name: Piece
   m_TagString: Untagged
@@ -105,22 +104,6 @@ BoxCollider2D:
   serializedVersion: 2
   m_Size: {x: 190, y: 190}
   m_EdgeRadius: 0
---- !u!114 &5001380633490382158
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5001380633490382156}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 67232376cab3aaf4ea62d06a815ef2c8, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  owner: 1
-  type: 2
-  overlayImage: {fileID: 3219514861311018734}
-  collider: {fileID: 5001380633490382163}
 --- !u!1 &6159100201060309944
 GameObject:
   m_ObjectHideFlags: 0
