+++ Tatedrez-Game/Assets/Scripts/Board/Pieces.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 51613b24f70b5894b87e738a490ba85b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
