+++ Tatedrez-Game/Assets/Scripts/Board/Pieces/Bishop.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: a5577a048efa4ded87407913ea98ad63
+timeCreated: 1694949429
