+++ Tatedrez-Game/Assets/Scripts/Board/Pieces/Knight.cs
@@ -1,0 +1,22 @@
+using System;
+using UnityEngine;
+
+namespace Tatedrez.Board
+{
+    public class Knight : Piece
+    {
+        protected override void Start()
+        {
+            type = PieceType.Knight;
+            base.Start();
+        }
+        
+        public override bool IsValidMovement(Vector2Int origin, Vector2Int target, Cell[,] cells)
+        {
+            var distanceX = Math.Abs(target.x - origin.x);
+            var distanceY = Math.Abs(target.y - origin.y);
+            
+            return (distanceX == 2 && distanceY == 1) || (distanceX == 1 && distanceY == 2);
+        }
+    }
+}
