+++ Tatedrez-Game/Assets/Scripts/Board/Pieces/Knight.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8fb4a634606d27746929aa605bf232d7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
