+++ Tatedrez-Game/Assets/Scripts/Board/Pieces/Piece.cs
@@ -4,7 +4,7 @@ using UnityEngine.UI;
 
 namespace Tatedrez.Board
 {
-    public class Piece : MonoBehaviour
+    public abstract class Piece : MonoBehaviour
     {
         [Serializable]
         public enum PieceType
@@ -15,19 +15,22 @@ namespace Tatedrez.Board
         }
 
         [SerializeField] private PlayerId owner;
-        [SerializeField] private PieceType type;
         [SerializeField] private Image overlayImage;
         [SerializeField] private BoxCollider2D collider;
-
+        
         private int positionOffset = 2;
         private bool isPlaced = false;
         private Color originalOverlayColor;
         private Vector3 initialPosition;
 
+        protected PieceType type;
+        
         public PlayerId Owner => owner;
         public bool IsPlaced => isPlaced;
 
-        private void Start()
+        public abstract bool IsValidMovement(Vector2Int origin, Vector2Int target, Cell[,] cells);
+        
+        protected virtual void Start()
         {
             switch (owner)
             {
@@ -71,54 +74,7 @@ namespace Tatedrez.Board
             collider.enabled = !isPlaced;
         }
 
-        public bool IsValidMovement(Vector2Int origin, Vector2Int target, Cell[,] cells)
-        {
-            var distanceX = Math.Abs(target.x - origin.x);
-            var distanceY = Math.Abs(target.y - origin.y);
-            var movementVector = target - origin;
-            
-            switch (type)
-            {
-                case PieceType.Knight:
-                    return (distanceX == 2 && distanceY == 1) || (distanceX == 1 && distanceY == 2);
-                
-                case PieceType.Rook:
-                    var isHorizontalMovement = distanceX == 0 || distanceY == 0;
-                    if (!isHorizontalMovement) return false;
-
-                    if (movementVector.magnitude > 1)
-                    {
-                        foreach (var cell in cells)
-                        {
-                            if (CellObstructingPath(cell, origin, movementVector))
-                            {
-                                return false;
-                            }
-                        }
-                    }
-                    return true;
-                
-                case PieceType.Bishop:
-                    var isDiagonalMovement = distanceX == distanceY;
-                    if (!isDiagonalMovement) return false;
-
-                    if (movementVector.magnitude > 2)
-                    {
-                        foreach (var cell in cells)
-                        {
-                            if (CellObstructingPath(cell, origin, movementVector))
-                            {
-                                return false;
-                            }
-                        }
-                    }
-                    return true;
-                
-            }
-            return false;
-        }
-
-        private bool CellObstructingPath(Cell cell, Vector2Int origin, Vector2Int movementVector)
+        protected bool CellObstructingPath(Cell cell, Vector2Int origin, Vector2Int movementVector)
         {
             var movementVectorToCell = cell.Position - origin;
