+++ Tatedrez-Game/Assets/Scripts/Board/Pieces/Rook.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: a846c3607fa4463994ee18c934e3ec90
+timeCreated: 1694949422
